Final_Project/requirements.txt
psycopg2-binary
pandas
Flask-SQLAlchemy
snowflake-connector-python